move agent loop to its own file
