+++ src/rmplan/rmplan.ts
@@ -1,31 +1,18 @@
 #!/usr/bin/env bun
+import chalk from 'chalk';
 import clipboardy from 'clipboardy';
 import { Command } from 'commander';
 import os from 'os';
 import path from 'path';
-import yaml from 'yaml';
+import { error, log, warn } from '../logging.js';
 import { getInstructionsFromEditor } from '../rmfilter/instructions.js';
 import { getGitRoot, logSpawn, setDebug, setQuiet } from '../rmfilter/utils.js';
 import { findFilesCore, type RmfindOptions } from '../rmfind/core.js';
-import {
-  findPendingTask,
-  markStepDone,
-  prepareNextStep,
-  executePostApplyCommand,
-  extractMarkdownToYaml,
-} from './actions.js';
-import { convertMarkdownToYaml, findYamlStart, cleanupEolComments } from './cleanup.js';
+import { extractMarkdownToYaml, markStepDone, prepareNextStep } from './actions.js';
+import { rmplanAgent } from './agent.js';
+import { cleanupEolComments } from './cleanup.js';
 import { loadEffectiveConfig } from './configLoader.js';
-import { planSchema } from './planSchema.js';
 import { planPrompt } from './prompt.js';
-import { boldMarkdownHeaders, closeLogFile, error, log, openLogFile, warn } from '../logging.js';
-import { DEFAULT_RUN_MODEL, runStreamingPrompt } from '../common/run_and_apply.js';
-import { applyLlmEdits } from '../apply-llm-edits/apply.js';
-import chalk from 'chalk';
-import { createExecutor, executors } from './executors/index.js';
-import type { AgentCommandSharedOptions } from './executors/types.js';
-import { createRetryRequester } from '../apply-llm-edits/retry.js';
-import { rmplanConfigSchema } from './configSchema.js';
 
 const program = new Command();
 program.name('rmplan').description('Generate and execute task plans using LLMs');
@@ -349,187 +336,6 @@ program
   .option('--steps <steps>', 'Number of steps to execute')
   .option('--no-log', 'Do not log to file')
   .allowExcessArguments(true)
-  .action(async (planFile, options) => {
-    const globalCliOptions = program.opts();
-    const config = await loadEffectiveConfig(globalCliOptions.config);
-
-    const agentExecutionModel = options.model || config.models?.execution;
-
-    if (!options['no-log']) {
-      let lastDot = planFile.lastIndexOf('.');
-      let logFilePath = lastDot !== -1 ? planFile.slice(0, lastDot) : planFile;
-      logFilePath += '-agent-output.md';
-      openLogFile(logFilePath);
-    }
-
-    const sharedExecutorOptions: AgentCommandSharedOptions = {
-      planFile,
-      baseDir: await getGitRoot(),
-      model: agentExecutionModel,
-    };
-    const selectedExecutorName = options.executor;
-    const buildExecutorResult = createExecutor(
-      selectedExecutorName,
-      {},
-      sharedExecutorOptions,
-      config
-    );
-    if ('error' in buildExecutorResult) {
-      error(buildExecutorResult.error);
-      process.exit(1);
-    }
-    const { factory: executorFactory, executor } = buildExecutorResult;
-    log(`Using executor: ${executorFactory.name}`);
-
-    log('Starting agent to execute plan:', planFile);
-    try {
-      let hasError = false;
-
-      const maxSteps = options.steps ? parseInt(options.steps, 10) : Infinity;
-      let stepCount = 0;
-      while (stepCount < maxSteps) {
-        stepCount++;
-
-        const fileContent = await Bun.file(planFile).text();
-        let parsed;
-        try {
-          parsed = yaml.parse(fileContent);
-        } catch (err) {
-          error('Failed to parse YAML:', err);
-          process.exit(1);
-        }
-
-        const planResult = planSchema.safeParse(parsed);
-        if (!planResult.success) {
-          error('Validation errors:', JSON.stringify(planResult.error.issues, null, 2));
-          process.exit(1);
-        }
-
-        const planData = planResult.data;
-        const pendingTaskInfo = findPendingTask(planData);
-        if (!pendingTaskInfo) {
-          log('Plan complete!');
-          break;
-        }
-
-        log(
-          boldMarkdownHeaders(
-            `# Iteration ${stepCount}: Task ${pendingTaskInfo.taskIndex + 1}, Step ${pendingTaskInfo.stepIndex + 1}...`
-          )
-        );
-
-        const executorStepOptions = executor.prepareStepOptions?.() ?? {};
-        const stepPreparationResult = await prepareNextStep(config, planFile, {
-          previous: true,
-          ...executorStepOptions,
-          model: executorStepOptions.model || agentExecutionModel,
-          selectSteps: false,
-        }).catch((err) => {
-          error('Failed to prepare next step:', err);
-          hasError = true;
-          return null;
-        });
-
-        if (!stepPreparationResult) {
-          break;
-        }
-
-        const { promptFilePath, taskIndex, stepIndex, rmfilterArgs } = stepPreparationResult;
-
-        let contextContent: string;
-
-        if (executorStepOptions.rmfilter) {
-          if (!promptFilePath || !rmfilterArgs) {
-            error(
-              'Executor requires rmfilter, but no prompt file path or rmfilter args were generated by prepareNextStep.'
-            );
-            hasError = true;
-            break;
-          }
-          log(boldMarkdownHeaders('\n## Generating Context with rmfilter\n'));
-          const rmfilterOutputPath = promptFilePath.replace('.md', '.xml');
-          const proc = logSpawn(['rmfilter', '--output', rmfilterOutputPath, ...rmfilterArgs], {
-            stdio: ['inherit', 'inherit', 'inherit'],
-          });
-          const exitRes = await proc.exited;
-          if (exitRes !== 0) {
-            error(`rmfilter exited with code ${exitRes}`);
-            hasError = true;
-            break;
-          }
-          contextContent = await Bun.file(rmfilterOutputPath).text();
-          // Clean up rmfilter output path if needed, or handle in executor
-        } else {
-          log(boldMarkdownHeaders('\n## Using Direct Prompt as Context\n'));
-          contextContent = stepPreparationResult.prompt;
-        }
-
-        try {
-          log(boldMarkdownHeaders('\n## Execution\n'));
-          await executor.execute(contextContent);
-        } catch (err) {
-          error('Execution step failed:', err);
-          hasError = true;
-          break;
-        }
-
-        // ---> NEW: Execute Post-Apply Commands <---
-        if (config.postApplyCommands && config.postApplyCommands.length > 0) {
-          log(boldMarkdownHeaders('\n## Running Post-Apply Commands'));
-          for (const commandConfig of config.postApplyCommands) {
-            const commandSucceeded = await executePostApplyCommand(commandConfig);
-            if (!commandSucceeded) {
-              // Error logging is handled within executePostApplyCommand
-              error(`Agent stopping because required command "${commandConfig.title}" failed.`);
-              hasError = true;
-              break;
-            }
-          }
-          if (hasError) {
-            break;
-          }
-        }
-        // ---> END NEW SECTION <---
-        let markResult;
-        try {
-          log(boldMarkdownHeaders('\n## Marking done\n'));
-          markResult = await markStepDone(
-            planFile,
-            { steps: 1, commit: true },
-            { taskIndex, stepIndex }
-          );
-          log(`Marked step as done: ${markResult.message.split('\n')[0]}`);
-          if (markResult.planComplete) {
-            log('Plan fully completed!');
-            break;
-          }
-        } catch (err) {
-          error('Failed to mark step as done:', err);
-          hasError = true;
-          break;
-        } finally {
-          if (promptFilePath && executorStepOptions.rmfilter) {
-            // Only unlink if rmfilter was supposed to use it
-            try {
-              await Bun.file(promptFilePath).unlink();
-            } catch (e) {
-              warn('Warning: failed to clean up temp file:', promptFilePath);
-            }
-          }
-        }
-      }
-
-      await closeLogFile();
-
-      if (hasError) {
-        error('Agent stopped due to error.');
-        process.exit(1);
-      }
-    } catch (err) {
-      error('Unexpected error during agent execution:', err);
-      error('Agent stopped due to error.');
-      process.exit(1);
-    }
-  });
+  .action((planFile, options) => rmplanAgent(planFile, options, program.opts()));
 
 await program.parseAsync(process.argv);
